Save markup functions to use = syntax instead of (){ ... }
This should make it readable that these functions should do nothing else.
Unfortunately, javascript does not like the idea of overridable object getters
inside constructors. The constructor ends up executing first and then never
retrieves the markup if vals are used instead of funs. So no way of restricting
the markup() fun to a val.